lab-6

+++ lab-6/dir_and_files/task8.py
@@ -1,16 +1,17 @@
 import os
 
-
 def delete_file(path):
+
     if os.path.exists(path):
-        try:
+
+        if os.access(path, os.W_OK):
             os.remove(path)
-            print(f"The file '{path}' has been successfully deleted.")
-        except Exception as e:
-            print(f"An error occurred while deleting the file '{path}': {e}")
+            print(f"File '{path}' has been deleted.")
+        else:
+            print(f"Unable to delete file '{path}'.")
+
     else:
-        print(f"The file '{path}' does not exist.")
+        print(f"File '{path}' doesn't exist.")
 
-
-file_path = ""
+file_path = "/Users/Admin/DocumentsClone/Lab-python/lab-6/dir_and_files/files_8/delete_file.txt"
 delete_file(file_path)
